moved columnTasks from BoardPage to a seperate component
