+++ src/components/Task/Task.js
@@ -7,13 +7,13 @@ import Tooltip from "components/Tooltip/Tooltip";
 
 import { Draggable } from "react-beautiful-dnd";
 
-const Task = ({ taskId, name, index, columnIndex, tags, people, removeTask }) => {
+const Task = ({ taskId, name, index, tags, people, removeTask }) => {
 	const [, modalDispatch] = useContext(ModalContext);
 
 	const poepleAnchorElement = useRef();
 	const tagsAnchorElement = useRef();
 
-	const openTaskDetailsModal = (event) => {
+	const openTaskDetailsModal = () => {
 		modalDispatch({
 			type: "OPEN",
 			payload: {
@@ -25,42 +25,40 @@ const Task = ({ taskId, name, index, columnIndex, tags, people, removeTask }) =>
 
 	return (
 		<Draggable draggableId={taskId} index={index}>
-			{(provided, snapshot) => {
-				return (
-					<div
-						ref={provided.innerRef}
-						{...provided.draggableProps}
-						{...provided.dragHandleProps}
-						className={`task-card ${snapshot.isDragging ? "isDragging" : ""}`}
-						onClick={openTaskDetailsModal}
-						style={{ ...provided.draggableProps.style }}
-					>
-						<h3 className="task-title">{name}</h3>
-						<div className="card-bottom">
-							<div className="task-tags">
-								<div className="tags" ref={tagsAnchorElement}>
-									{tags &&
-										tags.map(({ color, id, name }) => (
-											<div key={id} className={`tag-mini ${color}`}></div>
-										))}
-								</div>
-							</div>
-							<Tooltip anchorEl={tagsAnchorElement} index={index}>
-								{tags && tags.map(({ id, name }) => <span key={id}>{name}</span>)}
-							</Tooltip>
-							<div className="task-people" ref={poepleAnchorElement}>
-								{people &&
-									people.map(({ id, imageURL }) => (
-										<Image key={id} classes={["avatar"]} imageURL={imageURL} />
+			{(provided, snapshot) => (
+				<div
+					ref={provided.innerRef}
+					{...provided.draggableProps}
+					{...provided.dragHandleProps}
+					className={`task-card ${snapshot.isDragging ? "isDragging" : ""}`}
+					onClick={openTaskDetailsModal}
+					style={{ ...provided.draggableProps.style }}
+				>
+					<h3 className="task-title">{name}</h3>
+					<div className="card-bottom">
+						<div className="task-tags">
+							<div className="tags" ref={tagsAnchorElement}>
+								{tags &&
+									tags.map(({ color, id, name }) => (
+										<div key={id} className={`tag-mini ${color}`}></div>
 									))}
 							</div>
-							<Tooltip anchorEl={poepleAnchorElement} index={index}>
-								{people && people.map(({ id, username }) => <span key={id}>{username}</span>)}
-							</Tooltip>
 						</div>
+						<Tooltip anchorEl={tagsAnchorElement} index={index}>
+							{tags && tags.map(({ id, name }) => <span key={id}>{name}</span>)}
+						</Tooltip>
+						<div className="task-people" ref={poepleAnchorElement}>
+							{people &&
+								people.map(({ id, imageURL }) => (
+									<Image key={id} classes={["avatar"]} imageURL={imageURL} />
+								))}
+						</div>
+						<Tooltip anchorEl={poepleAnchorElement} index={index}>
+							{people && people.map(({ id, username }) => <span key={id}>{username}</span>)}
+						</Tooltip>
 					</div>
-				);
-			}}
+				</div>
+			)}
 		</Draggable>
 	);
 };
