+++ src/pages/BoardPage/BoardPage.js
@@ -1,6 +1,4 @@
 import React, { useState, useContext } from "react";
-import TaskColumn from "components/Task/TaskColumn";
-import TaskColumnDraggable from "components/Task/DraggableColumn";
 import "./BoardPage.scss";
 import ExpandText from "components/ExpandText/ExpandText";
 import Button from "components/Button/Button";
@@ -9,16 +7,15 @@ import LocalOfferIcon from "@material-ui/icons/LocalOffer";
 import { ModalContext } from "context/ModalContext";
 import { BoardMembers, TagForm } from "modalForms";
 
-import { DragDropContext, Droppable } from "react-beautiful-dnd";
+import TaskBoard from "./TaskBoard";
+
+import { DragDropContext } from "react-beautiful-dnd";
 
 import { boardTasks_DATA, taskColumns_DATA, boardInfo_DATA, boardTasks_2_DATA } from "data";
 
 const onDragEnd = (result, tasks, setTasks) => {
 	if (!result.destination) return;
-	console.log(result);
 	const { source, destination, type } = result;
-
-	console.log(result);
 
 	if (type === "droppableTaskToColumn") {
 		const indexOfSourceColumn = tasks.findIndex(({ id }) => id === source.droppableId);
@@ -44,15 +41,11 @@ const onDragEnd = (result, tasks, setTasks) => {
 };
 
 const BoardPage = ({ boardId }) => {
-	const [newColumn, setNewColumn] = useState("");
-
 	const [boardInfo] = useState({
 		name: boardInfo_DATA.name,
 		description: boardInfo_DATA.description,
 	});
 	const [tasks, setTasks] = useState(boardTasks_2_DATA);
-
-	const [columns, setColumns] = useState(taskColumns_DATA);
 
 	const [, modalDispatch] = useContext(ModalContext);
 
@@ -67,40 +60,6 @@ const BoardPage = ({ boardId }) => {
 			type: "OPEN",
 			payload: { render: <TagForm />, title: "Board Tags" },
 		});
-	};
-	const handleNewColumnChange = (event) => {
-		setNewColumn(event.target.value);
-	};
-	const createNewColumn = (event) => {
-		if (event.key === "Enter" && newColumn !== "") {
-			setNewColumn("");
-			const submittedColumn = {
-				id: newColumn,
-				name: newColumn,
-				tasks: [],
-			};
-			setTasks((tasks) => {
-				const tempTasks = [...tasks];
-				tempTasks.push(submittedColumn);
-				return tempTasks;
-			});
-		}
-	};
-	const removeColum = (columnIndex) => {
-		setColumns((columns) => {
-			const tempColumns = [...columns];
-			tempColumns.splice(columnIndex, 1);
-			return tempColumns;
-		});
-	};
-	const removeTask = (taskId) => {
-		setTasks((tasks) => {
-			const tempTasks = [...tasks];
-			const foundIndexOfTask = tempTasks.findIndex(({ id }) => id === taskId);
-			tempTasks.splice(foundIndexOfTask, 1);
-			return tempTasks;
-		});
-		modalDispatch({ type: "CLOSE" });
 	};
 
 	return (
@@ -119,38 +78,7 @@ const BoardPage = ({ boardId }) => {
 				</Button>
 			</div>
 			<DragDropContext onDragEnd={(result) => onDragEnd(result, tasks, setTasks)}>
-				<Droppable droppableId="droppable" type="droppableColumn" direction="horizontal">
-					{(provided, snapshot) => {
-						return (
-							<div className="board-page-container" ref={provided.innerRef}>
-								{tasks.map(({ id, name, tasks }, index) => (
-									<span key={id}>
-										<TaskColumnDraggable
-											columnId={id}
-											columnIndex={index}
-											removeTask={removeTask}
-											removeColumn={() => removeColum(index)}
-											columnName={name}
-											listOfTasks={tasks}
-										/>
-									</span>
-								))}
-								{provided.placeholder}
-							</div>
-						);
-					}}
-				</Droppable>
-				<div>
-					<div className="add-new-column">
-						<input
-							onKeyDown={createNewColumn}
-							value={newColumn}
-							onChange={handleNewColumnChange}
-							type="text"
-							placeholder="+ new column"
-						/>
-					</div>
-				</div>
+				<TaskBoard tasks={tasks} setTasks={setTasks} />
 			</DragDropContext>
 		</div>
 	);
