+++ src/pages/BoardPage/BoardPage.scss
@@ -11,40 +11,41 @@
 		}
 	}
 	.add-new-column {
-    display: flex;
-    flex-direction: column;
-    background-color: $blue-primary-overlay;
-    border-radius: 10px;
-    margin: 0 0.5rem;
-    padding: 0.2rem 0.5rem;
-    font-family: "Secular One", sans-serif;
-    width: 250px;
-    height: 30px;
-    align-items: center;
-    justify-content: center;
-    position: relative;
-    border: 1px darken($blue-primary-overlay, 40%) dashed;
+		display: flex;
+		flex-direction: column;
+		background-color: $blue-primary-overlay;
+		border-radius: 10px;
+		margin: 0 0.5rem;
+		padding: 0.2rem 0.5rem;
+		font-family: "Secular One", sans-serif;
+		width: 250px;
+		height: 30px;
+		align-items: center;
+		justify-content: center;
+		position: relative;
+		border: 1px darken($blue-primary-overlay, 40%) dashed;
 
-    input {
-      outline: none;
-      height: 100%;
-      width: 100%;
-      background: transparent;
-      border: none;
-      font-family: "Secular One", sans-serif;
-      margin: 0 0.5rem;
-      box-sizing: border-box;
-      &::placeholder {
-        color: darken($blue-primary-overlay, 30%);
-      } 
-    }
+		input {
+			outline: none;
+			height: 100%;
+			width: 100%;
+			background: transparent;
+			border: none;
+			font-family: "Secular One", sans-serif;
+			margin: 0 0.5rem;
+			box-sizing: border-box;
+			&::placeholder {
+				color: darken($blue-primary-overlay, 30%);
+			}
+		}
 	}
 	.board-page-container {
 		display: flex;
 		flex-direction: row;
 		overflow-x: scroll;
 		padding: 1rem 0;
-    align-items: flex-start;
+		align-items: flex-start;
+
 		&::-webkit-scrollbar {
 			height: 0.5em;
 		}
@@ -57,6 +58,13 @@
 		&::-webkit-scrollbar-thumb {
 			background-color: darken($gray-1, 15%);
 			border-radius: 10px;
+		}
+
+		.board-page-flex {
+			display: flex;
+			flex-direction: row;
+			align-items: flex-start;
+
 		}
 	}
 	.board-button-group {
